# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "RENTAL DIGITAL- API",
    "description": "API DESENVOLVIDA PARA GESTÃO DE ALUGUEL DE CARROS"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/cars": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "category": {
                    "example": "any"
                  },
                  "seats": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  },
                  "transmission": {
                    "example": "any"
                  },
                  "fuel": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  },
                  "available": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cars/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "category": {
                    "example": "any"
                  },
                  "seats": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  },
                  "transmission": {
                    "example": "any"
                  },
                  "fuel": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  },
                  "available": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/send-email": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "someBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhon Doe"
          },
          "age": {
            "type": "number",
            "example": 29
          },
          "about": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "name",
          "age"
        ]
      },
      "someResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhon Doe"
          },
          "age": {
            "type": "number",
            "example": 29
          },
          "diplomas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "school": {
                  "type": "string",
                  "example": "XYZ University"
                },
                "year": {
                  "type": "number",
                  "example": 2020
                },
                "completed": {
                  "type": "boolean",
                  "example": true
                },
                "internship": {
                  "type": "object",
                  "properties": {
                    "hours": {
                      "type": "number",
                      "example": 290
                    },
                    "location": {
                      "type": "string",
                      "example": "XYZ Company"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "someEnum": {
        "type": "string",
        "enum": [
          "red",
          "yellow",
          "green"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}